requirements.txt
praw==7.8.1
pandas==2.0.2
numpy==1.24.3
textblob==0.17.1
matplotlib==3.7.1
seaborn==0.12.2
pyyaml==6.0.1
scipy==1.11.2
nltk==3.8.1
wordcloud==1.9.3
tqdm==4.66.1
emoji==2.8.0
contractions==0.1.73
spacy>=3.7.2
gensim==4.3.2
torch==2.1.0
transformers==4.35.0
textstat==0.7.3
vaderSentiment==3.3.2
scikit-learn>=1.3.0
apify-client==1.4.0 